Check if alloc has been released before releasing on timeout
Changes after code review
